Move hooks to components directory Keeping all .tsx files in the app directory
