+++ src/app/(routes)/(auth)/login/login-form.spec.tsx
@@ -19,7 +19,7 @@ let mockHandleFieldChange = jest.fn();
 let mockValidateFields = jest.fn();
 
 // See https://stackoverflow.com/questions/65270255/mocking-react-custom-hook-with-jest
-jest.mock('../../../../utils/hooks/use-form-validation', () => ({
+jest.mock('../../../components/hooks/use-form-validation', () => ({
   __esModule: true,
   default: () => ({
     fieldState: mockFieldState,
@@ -29,7 +29,7 @@ jest.mock('../../../../utils/hooks/use-form-validation', () => ({
 }));
 
 const mockLogin = jest.fn();
-jest.mock('../../../../utils/hooks/use-auth', () => ({
+jest.mock('../../../components/hooks/use-auth', () => ({
   __esModule: true,
   default: () => ({
     login: mockLogin
